frontend: helpers: Move recentClusters stuff to own module
