+++ frontend/src/helpers/recentClusters.ts
@@ -1,0 +1,29 @@
+import { Cluster } from '../lib/k8s/cluster';
+
+const recentClustersStorageKey = 'recent_clusters';
+/**
+ * Adds the cluster name to the list of recent clusters in localStorage.
+ *
+ * @param cluster - the cluster to add to the list of recent clusters. Can be the name, or a Cluster object.
+ * @returns void
+ */
+export function setRecentCluster(cluster: string | Cluster) {
+  const recentClusters = getRecentClusters();
+  const clusterName = typeof cluster === 'string' ? cluster : cluster.name;
+  const currentClusters = recentClusters.filter(name => name !== clusterName);
+  const newClusters = [clusterName, ...currentClusters].slice(0, 3);
+  localStorage.setItem(recentClustersStorageKey, JSON.stringify(newClusters));
+}
+/**
+ * @returns the list of recent clusters from localStorage.
+ */
+export function getRecentClusters() {
+  const currentClustersStr = localStorage.getItem(recentClustersStorageKey) || '[]';
+  const recentClusters = JSON.parse(currentClustersStr) as string[];
+
+  if (!Array.isArray(recentClusters)) {
+    return [];
+  }
+
+  return recentClusters;
+}
